+++ tests/Unit/SubPageList/Lister/PageTest.php
@@ -16,10 +16,10 @@ class PageTest extends \PHPUnit_Framework_TestCase {
 
 	public function testConstructSetsFields() {
 		$title = $this->getMock( 'Title' );
-		$children = array(
+		$children = [
 			new Page( $this->getMock( 'Title' ) ),
 			new Page( $this->getMock( 'Title' ) )
-		);
+		];
 
 		$page = new Page( $title, $children );