+++ tests/Unit/SubPageList/Lister/UI/TreeListRendererTest.php
@@ -24,18 +24,18 @@ class TreeListRendererTest extends \PHPUnit_Framework_TestCase {
 		$this->assertRendersHierarchy(
 			new Page( Title::newFromText( $titleText ) ),
 			'',
-			array(
+			[
 				TreeListRenderer::OPT_SHOW_TOP_PAGE => false
-			)
+			]
 		);
 	}
 
 	public function titleTextProvider() {
-		return array(
-			array( 'A' ),
-			array( 'AAAAAAAA' ),
-			array( 'AA BB CC' ),
-		);
+		return [
+			[ 'A' ],
+			[ 'AAAAAAAA' ],
+			[ 'AA BB CC' ],
+		];
 	}
 
 	/**
@@ -48,7 +48,7 @@ class TreeListRendererTest extends \PHPUnit_Framework_TestCase {
 		);
 	}
 
-	private function assertRendersHierarchy( Page $page, $expected, array $options = array() ) {
+	private function assertRendersHierarchy( Page $page, $expected, array $options = [] ) {
 		$listRender = $this->newListRenderer( $options );
 
 		$actual = $listRender->renderHierarchy( $page );
@@ -78,10 +78,10 @@ class TreeListRendererTest extends \PHPUnit_Framework_TestCase {
 		$this->assertRendersHierarchy(
 			new Page(
 				Title::newFromText( 'AAA' ),
-				array(
+				[
 					new Page( Title::newFromText( 'CCC' ) ),
 					new Page( Title::newFromText( 'BBB' ) ),
-				)
+				]
 			),
 			'AAA
 * BBB
@@ -93,17 +93,17 @@ class TreeListRendererTest extends \PHPUnit_Framework_TestCase {
 		$this->assertRendersHierarchy(
 			new Page(
 				Title::newFromText( 'AAA' ),
-				array(
+				[
 					new Page( Title::newFromText( 'CCC' ) ),
 					new Page( Title::newFromText( 'BBB' ) ),
 					new Page( Title::newFromText( 'DDD' ) ),
-				)
+				]
 			),
 			'AAA
 * DDD
 * CCC
 * BBB',
-			array( 'sort' => 'desc' )
+			[ 'sort' => 'desc' ]
 		);
 	}
 
@@ -111,17 +111,17 @@ class TreeListRendererTest extends \PHPUnit_Framework_TestCase {
 		$this->assertRendersHierarchy(
 			new Page(
 				Title::newFromText( 'AAA' ),
-				array(
+				[
 					new Page( Title::newFromText( 'CCC' ) ),
 					new Page( Title::newFromText( 'BBB' ) ),
-				)
+				]
 			),
 			'AAA
 # BBB
 # CCC',
-			array(
+			[
 				TreeListRenderer::OPT_FORMAT => TreeListRenderer::FORMAT_OL
-			)
+			]
 		);
 	}
 
@@ -129,25 +129,25 @@ class TreeListRendererTest extends \PHPUnit_Framework_TestCase {
 		$this->assertRendersHierarchy(
 			new Page(
 				Title::newFromText( 'AAA' ),
-				array(
+				[
 					new Page(
 						Title::newFromText( 'BBB' ),
-						array(
+						[
 							new Page( Title::newFromText( '111' ) ),
 							new Page( Title::newFromText( '222' ) ),
-						)
+						]
 					),
 					new Page(
 						Title::newFromText( 'CCC' ),
-						array()
+						[]
 					),
 					new Page(
 						Title::newFromText( 'DDD' ),
-						array(
+						[
 							new Page( Title::newFromText( '333' ) ),
-						)
+						]
 					),
-				)
+				]
 			),
 			'AAA
 * BBB
@@ -163,33 +163,33 @@ class TreeListRendererTest extends \PHPUnit_Framework_TestCase {
 		$this->assertRendersHierarchy(
 			new Page(
 				Title::newFromText( 'AAA' ),
-				array(
+				[
 					new Page(
 						Title::newFromText( 'BBB' ),
-						array(
+						[
 							new Page( Title::newFromText( '111' ) ),
 							new Page( Title::newFromText( '222' ) ),
-						)
+						]
 					),
 					new Page(
 						Title::newFromText( 'CCC' ),
-						array()
+						[]
 					),
 					new Page(
 						Title::newFromText( 'DDD' ),
-						array(
+						[
 							new Page( Title::newFromText( '333' ) ),
-						)
+						]
 					),
-				)
+				]
 			),
 			'AAA
 * BBB
 * CCC
 * DDD',
-			array(
+			[
 				TreeListRenderer::OPT_MAX_DEPTH => 1
-			)
+			]
 		);
 	}